fts: co-locate reader tests into their modules
The reader/ split left the whole ~2,100-line test module in core.rs
while the code it exercised moved into cursor/phrase/sink/filter/work/
scorers/count/search/metadata. Relocate each test beside the code it
covers: search entry points to search.rs, union/AND scorer agreement
to scorers.rs, unranked match/count and term-df to count.rs, exclude
filters to filter.rs, phrase matching to phrase.rs, config/metadata to
metadata.rs, the top-k drain to sink.rs; core.rs keeps open/format/
fetch/iter/top-k. Shared blob fixtures move to a new test_util module.

This also clears the stale top-of-file imports the moves left behind:
with the tests co-located, each module's imports serve only its own
code and tests, so the reader/ split now compiles clean under
clippy --all-targets. No test dropped (74 reader tests, verified by
name) and no behavior change.
