+++ src/superfile/fts/reader/core.rs
@@ -16,7 +16,6 @@
 //! then it's a borrowed view).
 
 use std::{
-    cmp::Ordering,
     collections::{BinaryHeap, HashMap},
     ops::Range,
     sync::Arc,
@@ -24,57 +23,38 @@ use std::{
 
 use bytes::Bytes;
 
-use super::cursor::{BlockMeta, TermCursor, TermMeta};
-use super::filter::{AtomExcludeFilter, ExcludeFilter};
-use super::metadata::{ColumnMeta, FtsColumnConfig, NormTable, OpenOptions, default_tokenizer};
-use super::options::{Bm25SearchOptions, Bm25Stats, BoolMode};
+use super::cursor::{TermCursor, TermMeta};
+use super::filter::ExcludeFilter;
+use super::metadata::{ColumnMeta, FtsColumnConfig, NormTable, OpenOptions};
 use super::phrase::{AnyCursor, PhraseCursor};
-use super::sink::{
-    AndSink, CollectSink, CountSink, MustShouldSink, ScoreSink, TopKEntry, and_heap_push,
-    drain_top_k_desc,
-};
-use super::work::{
-    MatchWork, atom_cursor_bytes, atom_planned_ranges, term_cursor_bytes, term_cursor_ranges,
-};
-use serde::Deserialize;
+use super::sink::{TopKEntry, drain_top_k_desc};
+use super::work::{term_cursor_bytes, term_cursor_ranges};
 
-use crate::{
-    runtime_metrics::{
-        cpu::{thread_cpu_delta_ns, thread_cpu_ns},
-        op_stats::{metering_active, timed_section},
-    },
-    superfile::{
-        ReadError,
-        error::FtsError,
-        format::{
-            self, FST_SEPARATOR,
-            checksum::crc32c,
-            fts::{
-                HEADER_SIZE_V1_LEGACY as FTS_HEADER_SIZE, MAGIC_BYTES, U32_BYTES, U64_BYTES, hdr,
-                skip_entry, term_meta,
-            },
-        },
+use crate::superfile::{
+    ReadError,
+    error::FtsError,
+    format::{
+        self, FST_SEPARATOR,
+        checksum::crc32c,
         fts::{
-            bm25,
-            builder::{
-                DOC_LENGTHS_ENTRY_SIZE, SKIP_ENTRY_SIZE, TERM_META_POSITIONAL_SIZE, TERM_META_SIZE,
-            },
-            dict::{DictReader, make_key},
-            fst_value::FstValue,
-            positions::{decode_run, skip_run},
-            posting::{BLOCK_LEN, decode_block},
-            tokenize::{Tokenizer, tokenizer_for_name},
+            HEADER_SIZE_V1_LEGACY as FTS_HEADER_SIZE, MAGIC_BYTES, U32_BYTES, U64_BYTES, hdr,
+            term_meta,
         },
-        lazy_source::{LazyByteSource, PrefetchedSource, RangeCoalescePlan, Source},
     },
+    fts::{
+        builder::{DOC_LENGTHS_ENTRY_SIZE, TERM_META_SIZE},
+        dict::{DictReader, make_key},
+        fst_value::FstValue,
+        positions::decode_run,
+        tokenize::{Tokenizer, tokenizer_for_name},
+    },
+    lazy_source::{LazyByteSource, PrefetchedSource, RangeCoalescePlan, Source},
 };
 
 /// Largest gap worth overfetching when adjacent term postings share a request.
 const TERM_RANGE_COALESCE_MAX_GAP: usize = 64 * 1024;
 /// Maximum total gap bytes tolerated in one coalesced postings request.
 const TERM_RANGE_COALESCE_MAX_OVERFETCH: usize = 512 * 1024;
-
-/// Boolean-mode for multi-term queries.
 
 /// Per-term global BM25 idf (the raw `idf`, not `idf × (k1+1)`) keyed
 /// by term, used by [`Bm25Stats::Global`]. A term absent from the map
@@ -1578,28 +1558,15 @@ fn header_postings_length(header: &[u8]) -> Result<usize, FtsError> {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
-    use std::{collections::HashSet, sync::Arc};
-
+    use super::super::test_util::*;
     use super::*;
-    use crate::superfile::{
-        BytesLazyByteSource,
-        fts::{builder::FtsBuilder, tokenize::AsciiLowerTokenizer},
-    };
-
-    fn build_blob() -> (Bytes, String) {
-        // 3 docs, 1 column.
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false)
-            .expect("register column");
-        b.add_doc(0, 0, "rust async runtime").expect("add doc");
-        b.add_doc(0, 1, "tokio is a rust runtime").expect("add doc");
-        b.add_doc(0, 2, "java spring boot").expect("add doc");
-        let bytes = b.finish().expect("finish");
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        (Bytes::from(bytes), json.to_string())
-    }
+    use crate::superfile::BytesLazyByteSource;
+    use crate::superfile::fts::builder::FtsBuilder;
+    use crate::superfile::fts::reader::BoolMode;
+    use crate::superfile::fts::tokenize::AsciiLowerTokenizer;
+    use std::collections::HashSet;
 
     #[test]
     fn open_accepts_valid_blob() {
@@ -1840,371 +1807,6 @@ mod tests {
         ));
     }
 
-    #[tokio::test]
-    async fn search_returns_exact_doc_ids_for_known_term() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["rust"], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        // "rust" appears in doc 0 and doc 1.
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert!(ids.contains(&0), "doc 0 should match");
-        assert!(ids.contains(&1), "doc 1 should match");
-        assert!(!ids.contains(&2), "doc 2 should not match");
-    }
-
-    #[tokio::test]
-    async fn token_match_or_unions_and_intersects_unranked() {
-        // build_blob: doc0 "rust async runtime", doc1 "tokio is a rust
-        // runtime", doc2 "java spring boot".
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-
-        // Single token → its posting list, ascending.
-        assert_eq!(
-            r.token_match("body", &["rust"], BoolMode::Or)
-                .await
-                .expect("single")
-                .0,
-            vec![0, 1]
-        );
-        // OR = union (rust ∪ java).
-        assert_eq!(
-            r.token_match("body", &["rust", "java"], BoolMode::Or)
-                .await
-                .expect("or")
-                .0,
-            vec![0, 1, 2]
-        );
-        // AND = intersection (rust ∩ runtime).
-        assert_eq!(
-            r.token_match("body", &["rust", "runtime"], BoolMode::And)
-                .await
-                .expect("and")
-                .0,
-            vec![0, 1]
-        );
-        // AND with an absent token → empty.
-        assert!(
-            r.token_match("body", &["rust", "zzz"], BoolMode::And)
-                .await
-                .expect("and absent")
-                .0
-                .is_empty()
-        );
-        // OR ignores an absent token.
-        assert_eq!(
-            r.token_match("body", &["java", "zzz"], BoolMode::Or)
-                .await
-                .expect("or absent")
-                .0,
-            vec![2]
-        );
-        // Empty token list → empty.
-        assert!(
-            r.token_match("body", &[], BoolMode::And)
-                .await
-                .expect("empty")
-                .0
-                .is_empty()
-        );
-    }
-
-    #[tokio::test]
-    async fn token_match_count_matches_token_match_len() {
-        // The counting path (CountSink for AND, or_count_unranked for OR)
-        // must agree with token_match's materialized length on every
-        // shape — single token, OR union, AND intersection, absent
-        // tokens, and the empty list.
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let cases: &[(&[&str], BoolMode)] = &[
-            (&["rust"], BoolMode::Or),
-            (&["rust", "java"], BoolMode::Or),
-            (&["rust", "runtime"], BoolMode::And),
-            (&["rust", "zzz"], BoolMode::And),
-            (&["java", "zzz"], BoolMode::Or),
-            (&[], BoolMode::And),
-        ];
-        for (tokens, mode) in cases {
-            let len = r
-                .token_match("body", tokens, *mode)
-                .await
-                .expect("token_match")
-                .0
-                .len() as u64;
-            let count = r
-                .token_match_count("body", tokens, *mode)
-                .await
-                .expect("token_match_count")
-                .0;
-            assert_eq!(count, len, "count vs len for {tokens:?} {mode:?}");
-        }
-    }
-
-    #[tokio::test]
-    async fn or_count_spans_multiple_windows() {
-        // The windowed disjunction count must equal the union's true
-        // cardinality when the doc-id space spans several OR_WINDOW
-        // windows — exercising cross-window accumulation, the per-window
-        // popcount + clear, and dedup of docs that match multiple terms
-        // within one window. The naive ascending merge (token_match
-        // length) is the reference. Tied to OR_WINDOW so it keeps crossing
-        // the boundary if the window size changes.
-        const N_DOCS: u32 = OR_WINDOW * 2 + 500;
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::from("alpha "); // every doc
-            if i % 2 == 0 {
-                text.push_str("beta ");
-            }
-            if i % 3 == 0 {
-                text.push_str("gamma ");
-            }
-            if i % 5 == 0 {
-                text.push_str("delta ");
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-
-        let shapes: &[&[&str]] = &[
-            &["alpha"],                           // every doc
-            &["beta", "gamma"],                   // overlap on docs % 6
-            &["alpha", "beta", "gamma", "delta"], // all overlapping
-            &["gamma", "zzz_absent"],             // one absent term
-        ];
-        for terms in shapes {
-            let merge_len = r
-                .token_match("body", terms, BoolMode::Or)
-                .await
-                .expect("token_match")
-                .0
-                .len() as u64;
-            let count = r
-                .token_match_count("body", terms, BoolMode::Or)
-                .await
-                .expect("token_match_count")
-                .0;
-            assert_eq!(
-                count, merge_len,
-                "windowed count vs merge len for {terms:?}"
-            );
-        }
-        // `alpha` is in every doc, so its union count is exactly N_DOCS —
-        // pins the absolute multi-window cardinality, not just agreement
-        // with the merge.
-        assert_eq!(
-            r.token_match_count("body", &["alpha"], BoolMode::Or)
-                .await
-                .expect("count")
-                .0,
-            N_DOCS as u64
-        );
-    }
-
-    #[tokio::test]
-    async fn token_match_doc_set_matches_bm25_for_same_terms() {
-        // token_match(Or) must return exactly the doc set bm25 ranks.
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let mut bm25: Vec<u32> = r
-            .search("body", &["rust", "java"], 10, BoolMode::Or)
-            .await
-            .expect("search")
-            .into_iter()
-            .map(|(d, _)| d)
-            .collect();
-        bm25.sort_unstable();
-        let boolean = r
-            .token_match("body", &["rust", "java"], BoolMode::Or)
-            .await
-            .expect("boolean")
-            .0;
-        assert_eq!(bm25, boolean, "boolean Or doc set == bm25 doc set");
-    }
-
-    #[tokio::test]
-    async fn exhaustive_and_bmm_agree_on_top_k() {
-        // Build a larger blob so multi-term OR queries are
-        // interesting (some docs have multiple terms, some have one).
-        // Both algorithms must return identical top-K (descending
-        // score, ascending doc_id tiebreak).
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false)
-            .expect("register column");
-        // 20 docs sprinkled with mixed term combinations.
-        let docs = [
-            "alpha",
-            "beta",
-            "gamma",
-            "alpha beta",
-            "alpha gamma",
-            "beta gamma",
-            "alpha beta gamma",
-            "delta",
-            "epsilon",
-            "alpha delta",
-            "beta epsilon",
-            "gamma delta",
-            "alpha beta delta",
-            "alpha epsilon gamma",
-            "delta epsilon",
-            "alpha alpha alpha",
-            "beta beta beta",
-            "gamma gamma",
-            "alpha beta gamma delta epsilon",
-            "epsilon",
-        ];
-        for (i, text) in docs.iter().enumerate() {
-            b.add_doc(0, i as u32, text).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-
-        // Three terms with similar UBs — the heuristic should pick
-        // exhaustive for this shape, but we cross-check by calling
-        // both paths directly via the bench harness.
-        let terms: &[&str] = &["alpha", "beta", "gamma"];
-        let bmm = r
-            .search_with_algo_for_bench("body", terms, 5, OrAlgo::Bmm)
-            .await
-            .expect("bmm");
-        let exh = r
-            .search_with_algo_for_bench("body", terms, 5, OrAlgo::Exhaustive)
-            .await
-            .expect("exhaustive");
-        assert_eq!(bmm.len(), exh.len(), "result length mismatch");
-        for ((d_bmm, s_bmm), (d_exh, s_exh)) in bmm.iter().zip(exh.iter()) {
-            assert_eq!(d_bmm, d_exh, "doc_id mismatch");
-            assert!(
-                (s_bmm - s_exh).abs() < 1e-4,
-                "score mismatch: bmm={s_bmm} exhaustive={s_exh}"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn search_missing_term_or_returns_empty() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["nonexistent"], 10, BoolMode::Or)
-            .await
-            .expect("search");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn search_and_short_circuits_on_missing_term() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["rust", "nonexistent"], 10, BoolMode::And)
-            .await
-            .expect("search");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn search_and_intersects_term_postings() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        // "rust AND runtime" — both in doc 0 and doc 1.
-        let hits = r
-            .search("body", &["rust", "runtime"], 10, BoolMode::And)
-            .await
-            .expect("search");
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert!(ids.contains(&0));
-        assert!(ids.contains(&1));
-        assert!(!ids.contains(&2));
-    }
-
-    #[tokio::test]
-    async fn search_unknown_column_errors() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let err = r
-            .search("title", &["rust"], 10, BoolMode::Or)
-            .await
-            .expect_err("expected error");
-        assert!(matches!(err, FtsError::UnknownColumn(_)));
-    }
-
-    #[tokio::test]
-    async fn search_empty_terms_returns_empty() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &[], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn search_zero_k_returns_empty() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["rust"], 0, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn search_results_sorted_by_score_desc() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["rust"], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        for w in hits.windows(2) {
-            assert!(w[0].1 >= w[1].1, "scores should be descending");
-        }
-    }
-
-    #[tokio::test]
-    async fn search_limits_to_k() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["rust"], 1, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        assert_eq!(hits.len(), 1);
-    }
-
-    /// Build a corpus that exercises both the df=1 inline-encoded
-    /// path and the df ≥ 2 PFOR path side-by-side.
-    fn build_mixed_df_blob() -> (Bytes, String) {
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false)
-            .expect("register column");
-        // `common`     → df = 3 (PFOR form)
-        // `rust`       → df = 2 (PFOR form)
-        // `uniqzero`  → df = 1 (inline form)
-        // `uniqtwo`  → df = 1 (inline form)
-        b.add_doc(0, 0, "common rust uniqzero").expect("add doc");
-        b.add_doc(0, 1, "common rust").expect("add doc");
-        b.add_doc(0, 2, "common uniqtwo").expect("add doc");
-        let bytes = b.finish().expect("finish");
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        (Bytes::from(bytes), json.to_string())
-    }
-
     #[test]
     fn df1_inline_form_flag_set_on_fst_value() {
         // Verify the FST values for df=1 terms have bit 0 set
@@ -2249,64 +1851,6 @@ mod tests {
             }
             FstValue::Pfor { .. } => panic!("expected inline form"),
         }
-    }
-
-    #[tokio::test]
-    async fn df1_single_term_search_returns_one_doc() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["uniqzero"], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        assert_eq!(hits.len(), 1, "df=1 term should return exactly one hit");
-        assert_eq!(hits[0].0, 0, "uniqzero lives in doc 0");
-        assert!(hits[0].1 > 0.0, "score must be positive");
-    }
-
-    #[tokio::test]
-    async fn df1_in_or_query_combines_with_df_ge_2() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["uniqtwo", "rust"], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        // uniqtwo → doc 2; rust → docs 0, 1.
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert!(ids.contains(&0));
-        assert!(ids.contains(&1));
-        assert!(ids.contains(&2));
-    }
-
-    #[tokio::test]
-    async fn df1_in_and_query_intersects_correctly() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        // uniqzero ∩ rust = {doc 0}.
-        let hits = r
-            .search("body", &["uniqzero", "rust"], 10, BoolMode::And)
-            .await
-            .expect("FTS search");
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert_eq!(ids, vec![0]);
-        // uniqzero ∩ uniqtwo = ∅ (different docs).
-        let hits = r
-            .search("body", &["uniqzero", "uniqtwo"], 10, BoolMode::And)
-            .await
-            .expect("FTS search");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn df1_missing_term_returns_empty() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open FtsReader");
-        let hits = r
-            .search("body", &["nonexistentunique"], 10, BoolMode::Or)
-            .await
-            .expect("FTS search");
-        assert!(hits.is_empty());
     }
 
     #[test]
@@ -2382,168 +1926,6 @@ mod tests {
         );
     }
 
-    // ── ExcludeFilter (negation gate) ─────────────────────────────────
-    // `build_blob` plants: "rust" in docs 0 and 1, "java" in doc 2.
-
-    /// Build an `ExcludeFilter` over `terms` from the planted blob.
-    async fn exclude_filter_for(reader: &FtsReader, terms: &[&str]) -> ExcludeFilter {
-        let column_id = reader.resolve_column_id("body").expect("column exists");
-        let cursors = reader
-            .build_term_cursors(column_id, terms, None)
-            .await
-            .expect("build cursors");
-        ExcludeFilter::new(cursors)
-    }
-
-    #[tokio::test]
-    async fn exclude_filter_rejects_docs_in_negated_list() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let mut f = exclude_filter_for(&r, &["rust"]).await;
-        // "rust" is in docs 0 and 1 → excluded; doc 2 survives.
-        assert!(!f.admits(0));
-        assert!(!f.admits(1));
-        assert!(f.admits(2));
-    }
-
-    #[tokio::test]
-    async fn exclude_filter_missing_term_excludes_nothing() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // A negated term absent from the dictionary yields no cursor, so
-        // the filter admits every doc.
-        let mut f = exclude_filter_for(&r, &["nonexistent"]).await;
-        assert!(f.admits(0));
-        assert!(f.admits(1));
-        assert!(f.admits(2));
-    }
-
-    #[tokio::test]
-    async fn exclude_filter_multiple_negated_terms() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // Negating "rust" (docs 0,1) and "java" (doc 2) excludes all
-        // three — a doc is dropped if it matches ANY negated term.
-        let mut f = exclude_filter_for(&r, &["rust", "java"]).await;
-        assert!(!f.admits(0));
-        assert!(!f.admits(1));
-        assert!(!f.admits(2));
-    }
-
-    #[tokio::test]
-    #[cfg(debug_assertions)]
-    #[should_panic(expected = "non-monotonic")]
-    async fn exclude_filter_panics_on_non_monotonic_feed() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let mut f = exclude_filter_for(&r, &["rust"]).await;
-        // Feed a descending doc-id: `skip_to` can't seek backwards, so
-        // the debug assertion catches the contract violation.
-        let _ = f.admits(1);
-        let _ = f.admits(0);
-    }
-
-    // ── Additional coverage ───────────────────────────────────────────
-
-    #[test]
-    fn open_with_verify_crc_off_succeeds() {
-        // The trusted-storage fast path skips the four CRC scans but must
-        // still produce a fully usable reader.
-        let (blob, json) = build_blob();
-        let r = FtsReader::open_with(blob, &json, OpenOptions { verify_crc: false })
-            .expect("open with crc off");
-        assert_eq!(r.n_docs(), 3);
-        assert_eq!(r.fts_columns().collect::<Vec<_>>(), vec!["body"]);
-    }
-
-    #[test]
-    fn open_with_object_store_options_matches_crc_off() {
-        // `for_object_store` is the named constructor for the crc-off
-        // OpenOptions the lazy/object-store path uses.
-        let opts = OpenOptions::for_object_store();
-        assert!(!opts.verify_crc);
-        let (blob, json) = build_blob();
-        FtsReader::open_with(blob, &json, opts).expect("open object-store options");
-    }
-
-    #[test]
-    fn default_open_options_verifies_crc() {
-        assert!(OpenOptions::default().verify_crc);
-    }
-
-    #[test]
-    fn default_tokenizer_helper_is_ascii_lower() {
-        assert_eq!(default_tokenizer(), "ascii_lower");
-    }
-
-    #[test]
-    fn fts_column_config_missing_tokenizer_defaults() {
-        // A column JSON without the optional `tokenizer` field decodes to
-        // the ascii_lower default (round-trips an old file written before
-        // the field existed).
-        let (blob, _) = build_blob();
-        let json = r#"[{"name":"body"}]"#;
-        let r = FtsReader::open(blob, json).expect("open with terse json");
-        let cfg = r.fts_columns_config().next().expect("one column");
-        assert_eq!(cfg.name, "body");
-    }
-
-    #[test]
-    fn fts_columns_config_exposes_per_column_metadata() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let cols: Vec<&ColumnMeta> = r.fts_columns_config().collect();
-        assert_eq!(cols.len(), 1);
-        assert_eq!(cols[0].name, "body");
-        // Three non-empty docs ⇒ a positive average doc length and a
-        // populated per-doc normalization table.
-        assert!(cols[0].avgdl > 0.0);
-        assert_eq!(cols[0].dl_norm_k1.len(), 3);
-    }
-
-    #[test]
-    fn norm_table_footprint_is_one_byte_per_doc() {
-        // Memory guard: the resident length-norm table must stay at one
-        // byte per doc (plus the fixed 256-entry decode LUT), not the
-        // 4-byte-per-doc `f32` table it replaced. Build enough
-        // varied-length docs that the per-doc term dominates the LUT.
-        const N: u32 = 5_000;
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false)
-            .expect("register column");
-        for d in 0..N {
-            // Lengths cycle 1..=40 tokens so norms span many buckets and
-            // the table isn't a degenerate single value.
-            let words = (d % 40) + 1;
-            let text: String = (0..words).map(|w| format!("t{}x{w} ", d % 97)).collect();
-            b.add_doc(0, d, text.trim()).expect("add doc");
-        }
-        let bytes = b.finish().expect("finish");
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(Bytes::from(bytes), json).expect("open");
-        let nt = &r.columns[0].dl_norm_k1;
-
-        let per_doc = nt.bytes.capacity(); // 1 byte/doc
-        let lut = std::mem::size_of_val(&*nt.lut); // 256 * 4 = 1 KiB
-        let m2_bytes = per_doc + lut;
-        let f32_baseline = N as usize * std::mem::size_of::<f32>();
-
-        assert_eq!(nt.bytes.len(), N as usize, "one bucket byte per doc");
-        assert_eq!(nt.lut.len(), 256, "fixed 256-entry decode table");
-        // The whole point: strictly smaller than the old f32 table, and
-        // asymptotically 4× smaller (per-doc term is 1 byte vs 4).
-        assert!(
-            m2_bytes < f32_baseline,
-            "norm table {m2_bytes} B not smaller than f32 baseline {f32_baseline} B"
-        );
-        assert_eq!(
-            per_doc * 4,
-            f32_baseline,
-            "per-doc term is exactly 4× smaller"
-        );
-    }
-
     #[test]
     fn iter_column_terms_lists_every_term_in_lex_order() {
         // build_blob plants the union of tokens across the 3 docs.
@@ -2593,314 +1975,6 @@ mod tests {
         );
     }
 
-    #[tokio::test]
-    async fn term_df_reports_document_frequency() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // common → df 3 (PFOR header read), rust → df 2 (PFOR),
-        // uniqzero → df 1 (inline FST value), absent → 0.
-        assert_eq!(r.term_df("body", "common").await.expect("df").0, 3);
-        assert_eq!(r.term_df("body", "rust").await.expect("df").0, 2);
-        assert_eq!(r.term_df("body", "uniqzero").await.expect("df").0, 1);
-        assert_eq!(r.term_df("body", "missing").await.expect("df").0, 0);
-    }
-
-    #[tokio::test]
-    async fn term_df_unknown_column_errors() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let err = r.term_df("nope", "rust").await.expect_err("error");
-        assert!(matches!(err, FtsError::UnknownColumn(_)));
-    }
-
-    #[tokio::test]
-    async fn term_dfs_matches_per_term_term_df() {
-        let (blob, json) = build_mixed_df_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // Interleave the FST value kinds — PFOR (df>1), absent, inline
-        // (df=1), PFOR, absent — so a slot-mapping bug in the batched
-        // path (which fetches only the PFOR headers, then scatters the
-        // results back) would surface as a mismatch here.
-        let tokens = ["rust", "missing", "uniqzero", "common", "absent2"];
-        let batched = r.term_dfs("body", &tokens).await.expect("term_dfs").0;
-        // Element-wise identical to resolving each token on its own.
-        let mut per_term = Vec::with_capacity(tokens.len());
-        for t in tokens {
-            per_term.push(r.term_df("body", t).await.expect("term_df").0);
-        }
-        assert_eq!(
-            batched, per_term,
-            "batched term_dfs must equal per-term term_df"
-        );
-        // …and matches the planted ground truth (common=3, rust=2,
-        // uniqzero=1 inline, absent tokens=0).
-        assert_eq!(batched, vec![2, 0, 1, 3, 0], "planted document frequencies");
-        // Empty input short-circuits to empty output (no dict open, no fetch).
-        assert!(r.term_dfs("body", &[]).await.expect("empty").0.is_empty());
-    }
-
-    // ---- phrase atoms ----
-
-    /// Corpus with controlled adjacency for "new york": docs 0, 2
-    /// match (doc 4 twice); docs 1, 3 contain both words but never
-    /// adjacent in order.
-    fn build_phrase_blob() -> (Bytes, &'static str) {
-        use crate::superfile::fts::builder::FtsBuilder;
-        let mut b = FtsBuilder::new(crate::test_helpers::default_tokenizer());
-        b.register_column("title".into(), true).expect("register");
-        let docs = [
-            "new york city",
-            "york new haven",
-            "the new york times",
-            "new haven york",
-            "new york new york",
-        ];
-        for (i, d) in docs.iter().enumerate() {
-            b.add_doc(0, i as u32, d).expect("add doc");
-        }
-        (
-            Bytes::from(b.finish().expect("finish")),
-            r#"[{"name":"title","tokenizer":"ascii_lower","positions":true}]"#,
-        )
-    }
-
-    fn phrase(terms: &[&str]) -> Vec<Vec<String>> {
-        vec![terms.iter().map(|t| t.to_string()).collect()]
-    }
-
-    #[tokio::test]
-    async fn phrase_matches_adjacent_in_order_only() {
-        let (blob, json) = build_phrase_blob();
-        let r = FtsReader::open(blob, json).expect("open");
-        let phrases = phrase(&["new", "york"]);
-        let hits = r
-            .search_excluding(
-                "title",
-                ClauseLists {
-                    should_phrases: &phrases,
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect("phrase search");
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        let mut sorted = ids.clone();
-        sorted.sort_unstable();
-        assert_eq!(sorted, vec![0, 2, 4], "adjacency in order only");
-        // Doc 4 has the phrase twice — highest tf, and with uniform
-        // doc lengths in play its score must strictly exceed doc 0's
-        // (same length, tf 1... doc 0 len 3, doc 4 len 4; tf=2 wins).
-        assert_eq!(hits[0].0, 4, "double occurrence ranks first");
-    }
-
-    #[tokio::test]
-    async fn phrase_composes_with_clauses() {
-        let (blob, json) = build_phrase_blob();
-        let r = FtsReader::open(blob, json).expect("open");
-        let ny = phrase(&["new", "york"]);
-
-        // Must-phrase + must-term: "the" only in doc 2.
-        let hits = r
-            .search_excluding(
-                "title",
-                ClauseLists {
-                    musts: &["the"],
-                    must_phrases: &ny,
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect("must phrase + term");
-        assert_eq!(
-            hits.iter().map(|(d, _)| *d).collect::<Vec<_>>(),
-            vec![2],
-            "+\"new york\" +the"
-        );
-
-        // Negated phrase: haven-docs minus the phrase docs.
-        let hits = r
-            .search_excluding(
-                "title",
-                ClauseLists {
-                    shoulds: &["haven"],
-                    negative_phrases: &ny,
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect("negated phrase");
-        let mut ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        ids.sort_unstable();
-        assert_eq!(ids, vec![1, 3], "haven docs don't contain the phrase");
-    }
-
-    #[tokio::test]
-    async fn phrase_with_absent_member_matches_nothing() {
-        let (blob, json) = build_phrase_blob();
-        let r = FtsReader::open(blob, json).expect("open");
-        let ghost = phrase(&["new", "zealand"]);
-        let hits = r
-            .search_excluding(
-                "title",
-                ClauseLists {
-                    must_phrases: &ghost,
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect("ghost phrase");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn phrase_on_positionless_column_is_typed_error() {
-        use crate::superfile::fts::builder::FtsBuilder;
-        let mut b = FtsBuilder::new(crate::test_helpers::default_tokenizer());
-        b.register_column("title".into(), false).expect("register");
-        b.add_doc(0, 0, "new york").expect("add doc");
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let r =
-            FtsReader::open(blob, r#"[{"name":"title","tokenizer":"ascii_lower"}]"#).expect("open");
-        let phrases = phrase(&["new", "york"]);
-        let err = r
-            .search_excluding(
-                "title",
-                ClauseLists {
-                    should_phrases: &phrases,
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect_err("must be a typed error");
-        assert!(matches!(err, FtsError::PositionsUnavailable { .. }));
-    }
-
-    #[tokio::test]
-    async fn search_excluding_drops_negated_docs() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // "runtime" hits docs 0 and 1; negate "async" (only in doc 0).
-        let hits = r
-            .search_excluding(
-                "body",
-                ClauseLists {
-                    shoulds: &["runtime"],
-                    negatives: &["async"],
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect("search excluding");
-        let ids: Vec<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert_eq!(ids, vec![1], "doc 0 excluded by negated 'async'");
-    }
-
-    #[tokio::test]
-    async fn search_excluding_negation_only_errors() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let err = r
-            .search_excluding(
-                "body",
-                ClauseLists {
-                    negatives: &["rust"],
-                    ..ClauseLists::default()
-                },
-                10,
-                f32::NEG_INFINITY,
-            )
-            .await
-            .expect_err("negation-only");
-        assert!(matches!(err, FtsError::NegationOnly));
-    }
-
-    #[tokio::test]
-    async fn search_excluding_no_terms_at_all_is_empty() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        let hits = r
-            .search_excluding("body", ClauseLists::default(), 10, f32::NEG_INFINITY)
-            .await
-            .expect("empty");
-        assert!(hits.is_empty());
-    }
-
-    #[tokio::test]
-    async fn search_with_floor_prunes_below_floor() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // An impossibly high floor prunes every doc.
-        let hits = r
-            .search_with_floor("body", &["rust"], 10, BoolMode::Or, 1e9)
-            .await
-            .expect("floored search");
-        assert!(hits.is_empty(), "floor above all scores prunes everything");
-    }
-
-    #[tokio::test]
-    async fn search_multi_weights_and_combines_columns() {
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("title".into(), false).expect("register");
-        b.register_column("body".into(), false).expect("register");
-        // doc 0: title "rust"; doc 1: body "rust"; doc 2: neither.
-        b.add_doc(0, 0, "rust").expect("add");
-        b.add_doc(1, 0, "systems").expect("add");
-        b.add_doc(0, 1, "python").expect("add");
-        b.add_doc(1, 1, "rust ml").expect("add");
-        b.add_doc(0, 2, "go").expect("add");
-        b.add_doc(1, 2, "concurrency").expect("add");
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"title","tokenizer":"ascii_lower"},{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let hits = r
-            .search_multi(&[("title", 1.0), ("body", 1.0)], "rust", 10, BoolMode::Or)
-            .await
-            .expect("multi");
-        let ids: HashSet<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert!(ids.contains(&0));
-        assert!(ids.contains(&1));
-        assert!(!ids.contains(&2));
-    }
-
-    #[tokio::test]
-    async fn search_or_range_restricts_to_doc_id_window() {
-        // Larger corpus so an OR query spans several doc ids and the
-        // ranged path actually clips some out.
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..8u32 {
-            b.add_doc(0, i, "alpha beta").expect("add");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        // Restrict to [2, 5): only docs 2,3,4 are eligible.
-        let hits = r
-            .search_or_range_pretokenized("body", &["alpha", "beta"], 100, 2, 5)
-            .await
-            .expect("ranged search");
-        let ids: HashSet<u32> = hits.iter().map(|(d, _)| *d).collect();
-        assert_eq!(
-            ids,
-            [2u32, 3, 4].into_iter().collect(),
-            "only docs in [2,5) returned"
-        );
-    }
-
     /// A scan's top-k heap is preallocated for the docs it can actually
     /// rank: the whole superfile un-ranged, the window's width when
     /// ranged. Sizing a ranged scan by `n_docs` is what made a sliced
@@ -2945,716 +2019,6 @@ mod tests {
         );
     }
 
-    /// Regression: the ranged OR entry must produce the same results as
-    /// the un-ranged path for ANY partition of the doc space, on BOTH of
-    /// the kernels its dispatch can now pick. Before the fix it hardcoded
-    /// MaxScore+BMM, so a query sliced into sub-ranges (the fan-out shape
-    /// a compacted table takes) ran a different kernel than the same query
-    /// un-ranged — uniform broad ORs degraded 11-24x post-compaction.
-    #[tokio::test]
-    async fn search_or_range_partitions_agree_with_unranged() {
-        /// Docs in the planted corpus — spans several 4096-doc OR windows
-        /// and many 128-doc posting blocks.
-        const N_DOCS: u32 = 6_000;
-        /// Ask for every match so partition union == full result set.
-        const K_ALL: usize = N_DOCS as usize;
-        /// Top-k size for the truncated comparison.
-        const K_TOP: usize = 10;
-
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            // Deterministic mixed-df corpus: four uniform terms with
-            // varying tf (windowed-union shape), plus one rare term
-            // (dominant-UB / BMM shape when queried with two commons).
-            let mut text = String::new();
-            for (t, name) in ["alpha", "beta", "gamma", "delta"].iter().enumerate() {
-                let h = i.wrapping_mul(31).wrapping_add(t as u32 * 17) % 5;
-                for _ in 0..h {
-                    text.push_str(name);
-                    text.push(' ');
-                }
-            }
-            if i % 2000 == 7 {
-                text.push_str("rareterm ");
-            }
-            if text.is_empty() {
-                text.push_str("filler");
-            }
-            b.add_doc(0, i, &text).expect("add");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-
-        // Uniform 4-term OR routes to the windowed union; the
-        // rare+common mix keeps a dominant term UB and stays on BMM.
-        // Assert the routing rather than assume it — a corpus tweak that
-        // silently stopped exercising one branch would otherwise turn
-        // this into a test of the other branch twice.
-        let shapes: [&[&str]; 2] = [
-            &["alpha", "beta", "gamma", "delta"],
-            &["rareterm", "alpha", "beta"],
-        ];
-        let column_id = r.resolve_column_id("body").expect("column");
-        let uniform_cursors = r
-            .build_term_cursors(column_id, shapes[0], None)
-            .await
-            .expect("cursors");
-        assert!(
-            prefer_windowed_union(&uniform_cursors),
-            "uniform shape must route to the windowed ranged branch"
-        );
-        let dominant_cursors = r
-            .build_term_cursors(column_id, shapes[1], None)
-            .await
-            .expect("cursors");
-        assert!(
-            !prefer_windowed_union(&dominant_cursors),
-            "dominant-UB shape must route to the BMM ranged branch"
-        );
-        // Uneven partitions, including window-boundary-crossing cuts.
-        let partitions: [&[(u32, u32)]; 3] = [
-            &[(0, N_DOCS)],
-            &[(0, 3_000), (3_000, N_DOCS)],
-            &[(0, 100), (100, 4_097), (4_097, 5_000), (5_000, N_DOCS)],
-        ];
-
-        for terms in shapes {
-            let full = r
-                .search("body", terms, K_ALL, BoolMode::Or)
-                .await
-                .expect("un-ranged search");
-            let mut full_sorted: Vec<(u32, u32)> =
-                full.iter().map(|&(d, s)| (d, s.to_bits())).collect();
-            full_sorted.sort_unstable();
-
-            for cuts in partitions {
-                let mut merged: Vec<(u32, f32)> = Vec::new();
-                for &(lo, hi) in cuts {
-                    merged.extend(
-                        r.search_or_range_pretokenized("body", terms, K_ALL, lo, hi)
-                            .await
-                            .expect("ranged search"),
-                    );
-                }
-                let mut merged_sorted: Vec<(u32, u32)> =
-                    merged.iter().map(|&(d, s)| (d, s.to_bits())).collect();
-                merged_sorted.sort_unstable();
-                assert_eq!(
-                    merged_sorted, full_sorted,
-                    "partition union must equal the un-ranged result \
-                     (terms={terms:?}, cuts={cuts:?})"
-                );
-
-                // Top-k contract: resorting the merged pool by
-                // (score desc, doc asc) reproduces the un-ranged top-k.
-                let mut pool = merged.clone();
-                pool.sort_unstable_by(|a, b| {
-                    b.1.partial_cmp(&a.1)
-                        .expect("BM25 scores are finite")
-                        .then(a.0.cmp(&b.0))
-                });
-                pool.truncate(K_TOP);
-                let top: Vec<(u32, u32)> = pool.iter().map(|&(d, s)| (d, s.to_bits())).collect();
-                let full_top: Vec<(u32, u32)> = full
-                    .iter()
-                    .take(K_TOP)
-                    .map(|&(d, s)| (d, s.to_bits()))
-                    .collect();
-                assert_eq!(
-                    top, full_top,
-                    "merged top-{K_TOP} must equal un-ranged top-{K_TOP} \
-                     (terms={terms:?}, cuts={cuts:?})"
-                );
-            }
-        }
-    }
-
-    /// The prebuilt-cursor ranged path must be byte-identical to fresh
-    /// per-call builds — it is the same search minus the redundant fetch
-    /// and parse, so any divergence is a sharing bug (walk state leaking
-    /// between clones, stale first-block decode, ...). One set serves
-    /// overlapping windows and a repeated window to force reuse.
-    #[tokio::test]
-    async fn search_or_range_prebuilt_matches_fresh_calls() {
-        /// Docs in the planted corpus (multiple OR windows and blocks).
-        const N_DOCS: u32 = 6_000;
-        /// Ask for every match so whole result sets are compared.
-        const K_ALL: usize = N_DOCS as usize;
-
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::new();
-            for (t, name) in ["alpha", "beta", "gamma", "delta"].iter().enumerate() {
-                let h = i.wrapping_mul(31).wrapping_add(t as u32 * 17) % 5;
-                for _ in 0..h {
-                    text.push_str(name);
-                    text.push(' ');
-                }
-            }
-            if text.is_empty() {
-                text.push_str("filler");
-            }
-            b.add_doc(0, i, &text).expect("add");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-
-        let terms: &[&str] = &["alpha", "beta", "gamma", "delta"];
-        let set = r
-            .build_or_cursor_set("body", terms, None)
-            .await
-            .expect("set");
-        let windows = [
-            (0u32, N_DOCS),
-            (0, 3_000),
-            (2_000, 4_097),
-            (3_000, N_DOCS),
-            (0, N_DOCS),
-        ];
-        for (lo, hi) in windows {
-            let fresh = r
-                .search_or_range_pretokenized("body", terms, K_ALL, lo, hi)
-                .await
-                .expect("fresh ranged search");
-            let pre = r
-                .search_or_range_prebuilt(&set, K_ALL, lo, hi, f32::NEG_INFINITY)
-                .expect("prebuilt ranged search");
-            let fresh_bits: Vec<(u32, u32)> =
-                fresh.iter().map(|&(d, s)| (d, s.to_bits())).collect();
-            let pre_bits: Vec<(u32, u32)> = pre.iter().map(|&(d, s)| (d, s.to_bits())).collect();
-            assert_eq!(pre_bits, fresh_bits, "window ({lo},{hi})");
-        }
-    }
-
-    #[tokio::test]
-    async fn search_or_range_degenerate_inputs_are_empty() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        // Empty terms, k == 0, and an inverted range all short-circuit.
-        assert!(
-            r.search_or_range_pretokenized("body", &[], 10, 0, 3)
-                .await
-                .expect("empty terms")
-                .is_empty()
-        );
-        assert!(
-            r.search_or_range_pretokenized("body", &["rust"], 0, 0, 3)
-                .await
-                .expect("zero k")
-                .is_empty()
-        );
-        assert!(
-            r.search_or_range_pretokenized("body", &["rust"], 10, 3, 3)
-                .await
-                .expect("empty range")
-                .is_empty()
-        );
-    }
-
-    #[tokio::test]
-    async fn search_or_range_with_floor_prunes() {
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..8u32 {
-            b.add_doc(0, i, "alpha beta").expect("add");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let hits = r
-            .search_or_range_pretokenized_with_floor(
-                "body",
-                &["alpha", "beta"],
-                100,
-                0,
-                8,
-                1e9,
-                None,
-            )
-            .await
-            .expect("floored ranged search");
-        assert!(hits.is_empty(), "floor above all scores prunes everything");
-    }
-
-    #[tokio::test]
-    async fn search_with_algo_wand_bmw_agrees_with_bmm() {
-        // The historical WAND+BMW baseline must agree with the production
-        // BMM path on the planted corpus.
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        let docs = [
-            "alpha beta",
-            "alpha",
-            "beta gamma",
-            "alpha beta gamma",
-            "gamma",
-            "alpha gamma",
-            "beta",
-            "alpha beta gamma",
-        ];
-        for (i, t) in docs.iter().enumerate() {
-            b.add_doc(0, i as u32, t).expect("add");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let terms: &[&str] = &["alpha", "beta", "gamma"];
-        let bmm = r
-            .search_with_algo_for_bench("body", terms, 5, OrAlgo::Bmm)
-            .await
-            .expect("bmm");
-        let wand = r
-            .search_with_algo_for_bench("body", terms, 5, OrAlgo::WandBmw)
-            .await
-            .expect("wand");
-        assert_eq!(bmm.len(), wand.len());
-        for ((db, sb), (dw, sw)) in bmm.iter().zip(wand.iter()) {
-            assert_eq!(db, dw, "doc_id mismatch");
-            assert!((sb - sw).abs() < 1e-4, "score mismatch {sb} vs {sw}");
-        }
-    }
-
-    #[tokio::test]
-    async fn wand_bmw_exercises_block_skips_on_multi_block_lists() {
-        // A corpus large enough that the common terms span several
-        // 128-doc posting blocks, with five query terms of differing
-        // document frequency and a handful of docs carrying all five.
-        // Running WAND+BMW at a small k forces the pivot to move, the
-        // block-upper-bound skip to fire, lagging cursors to re-align,
-        // and the 4-wide SIMD scoring pack to be used on the
-        // all-terms docs — then cross-checks the result against BMM.
-
-        /// Total planted docs; well over several `BLOCK_LEN` (128) so
-        /// the dense-term posting lists occupy multiple blocks.
-        const N_DOCS: u32 = 400;
-        /// Requested top-K — small, so the heap fills early and the
-        /// score threshold starts pruning blocks.
-        const K: usize = 5;
-
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::new();
-            // `alpha` in ~every doc, `beta` in ~half, `gamma` every
-            // 5th, `delta` every 13th, `epsilon` every 29th — a
-            // descending-df mix that makes the WAND pivot non-trivial.
-            text.push_str("alpha ");
-            if i % 2 == 0 {
-                text.push_str("beta ");
-            }
-            if i % 5 == 0 {
-                text.push_str("gamma ");
-            }
-            if i % 13 == 0 {
-                text.push_str("delta ");
-            }
-            if i % 29 == 0 {
-                text.push_str("epsilon ");
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-
-        let terms: &[&str] = &["alpha", "beta", "gamma", "delta", "epsilon"];
-        let wand = r
-            .search_with_algo_for_bench("body", terms, K, OrAlgo::WandBmw)
-            .await
-            .expect("wand");
-        let bmm = r
-            .search_with_algo_for_bench("body", terms, K, OrAlgo::Bmm)
-            .await
-            .expect("bmm");
-        assert_eq!(wand.len(), bmm.len(), "result length mismatch");
-        assert_eq!(wand.len(), K, "expected a full top-K");
-        for ((dw, sw), (db, sb)) in wand.iter().zip(bmm.iter()) {
-            assert_eq!(dw, db, "doc_id mismatch wand={dw} bmm={db}");
-            assert!((sw - sb).abs() < 1e-4, "score mismatch {sw} vs {sb}");
-        }
-    }
-
-    #[tokio::test]
-    async fn windowed_union_agrees_with_bmm() {
-        // The windowed union scorer must return the identical top-k as
-        // the production MaxScore+BMM path — across term counts, k values,
-        // and the uniform-UB (common-term) shape it targets. N_DOCS spans
-        // multiple windows (and many BLOCK_LEN=128 posting blocks), so the
-        // walk exercises the multi-window path: base advancing to the next
-        // window, empty-window skipping, and cross-window monotonicity —
-        // not just a single window. Tied to OR_WINDOW so it keeps crossing
-        // the boundary if the window size changes.
-        const N_DOCS: u32 = OR_WINDOW * 2 + 500;
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::from("alpha zeta eta theta "); // ~every doc
-            if i % 2 == 0 {
-                text.push_str("beta ");
-            }
-            if i % 3 == 0 {
-                text.push_str("gamma ");
-            }
-            if i % 5 == 0 {
-                text.push_str("delta ");
-            }
-            if i % 7 == 0 {
-                text.push_str("epsilon ");
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let col = r.resolve_column_id("body").expect("col");
-        let uniform_terms: &[&str] = &["zeta", "eta", "theta"];
-        let uniform_cursors = r
-            .build_term_cursors(col, uniform_terms, None)
-            .await
-            .expect("uniform cursors");
-        assert!(
-            prefer_windowed_union(&uniform_cursors),
-            "production router should select windowed union for equal upper bounds"
-        );
-
-        let shapes: &[&[&str]] = &[
-            &["alpha", "beta"],
-            &["alpha", "beta", "gamma"],
-            &["beta", "gamma", "delta"], // no single dominator
-            &["alpha", "beta", "gamma", "delta", "epsilon"],
-            uniform_terms,
-        ];
-        for terms in shapes {
-            for k in [1usize, 5, 50, 1000] {
-                let bmm = r
-                    .search_with_algo_for_bench("body", terms, k, OrAlgo::Bmm)
-                    .await
-                    .expect("bmm");
-                let win = r
-                    .search_with_algo_for_bench("body", terms, k, OrAlgo::Windowed)
-                    .await
-                    .expect("windowed");
-                assert_eq!(bmm.len(), win.len(), "len mismatch {terms:?} k={k}");
-                for ((db, sb), (dw, sw)) in bmm.iter().zip(win.iter()) {
-                    assert_eq!(db, dw, "doc_id mismatch {terms:?} k={k}: bmm={db} win={dw}");
-                    assert!(
-                        (sb - sw).abs() < 1e-4,
-                        "score mismatch {terms:?} k={k}: {sb} vs {sw}"
-                    );
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn wand_bmw_2term_no_floor_agrees_with_bmm() {
-        // The small-k 2-term production path (`run_wand_bmw`) must return
-        // the identical top-k as MaxScore+BMM on the same inputs, across k.
-        // It is only reached floor-free (the dispatcher routes to MaxScore
-        // when a cross-segment floor is live), so both sides run unfloored
-        // (`NEG_INFINITY`). Multi-window corpus so WAND exercises block
-        // skips; `gamma` rarer than `beta` rarer than `alpha`.
-        const N_DOCS: u32 = OR_WINDOW * 2 + 500;
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::from("alpha ");
-            if i % 2 == 0 {
-                text.push_str("beta ");
-            }
-            if i % 3 == 0 {
-                text.push_str("gamma ");
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let col = r.resolve_column_id("body").expect("col");
-
-        let shapes: &[&[&str]] = &[&["alpha", "beta"], &["beta", "gamma"], &["alpha", "gamma"]];
-        for terms in shapes {
-            for k in [1usize, 5, 50, 128] {
-                let cw = r
-                    .build_term_cursors(col, terms, None)
-                    .await
-                    .expect("cursors");
-                let cb = r
-                    .build_term_cursors(col, terms, None)
-                    .await
-                    .expect("cursors");
-                let wand = r.run_wand_bmw(col, cw, k).expect("wand");
-                let bmm = r
-                    .run_max_score_bmm(col, cb, k, None, f32::NEG_INFINITY)
-                    .expect("bmm");
-                assert_eq!(wand.len(), bmm.len(), "len mismatch {terms:?} k={k}");
-                for ((dw, sw), (db, sb)) in wand.iter().zip(bmm.iter()) {
-                    assert_eq!(dw, db, "doc mismatch {terms:?} k={k}: {dw} vs {db}");
-                    assert!(
-                        (sw - sb).abs() < 1e-4,
-                        "score mismatch {terms:?} k={k}: {sw} vs {sb}"
-                    );
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn two_term_rare_anchor_gates_on_df_ratio() {
-        // `df` is read onto the cursor, and the 2-term WAND router fires
-        // only when one posting list is >= WAND_BMW_2TERM_DF_RATIO× shorter
-        // than the other (a rare anchor), not when both terms are common.
-        const N_DOCS: u32 = 4000;
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::from("common "); // every doc
-            if i % 2 == 0 {
-                text.push_str("frequent "); // half the docs
-            }
-            if i % 200 == 0 {
-                text.push_str("rare "); // ~1/200 of docs
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let col = r.resolve_column_id("body").expect("col");
-
-        // common (df≈N) + rare (df≈N/200): ratio 200 ≥ 16 → anchor.
-        let anchored = r
-            .build_term_cursors(col, &["common", "rare"], None)
-            .await
-            .expect("cursors");
-        assert!(
-            two_term_has_rare_anchor(&anchored),
-            "rare+common should have a rare anchor"
-        );
-        // common (df≈N) + frequent (df≈N/2): ratio 2 < 16 → no anchor.
-        let uniform = r
-            .build_term_cursors(col, &["common", "frequent"], None)
-            .await
-            .expect("cursors");
-        assert!(
-            !two_term_has_rare_anchor(&uniform),
-            "two common terms should not anchor"
-        );
-    }
-
-    #[test]
-    fn deep_k_dominant_union_reroutes_only_when_list_is_long() {
-        // The deep-k reroute to the windowed scorer fires only when
-        // pruning is dead (k reaches the rarer terms' combined df) AND the
-        // dominant list is long enough to amortize the window setup.
-        const LONG: u64 = 3_000_000; // dominant common term
-        const RARE: u64 = 500; // rare second term
-        let total = LONG + RARE;
-
-        // Deep k (>= rest_df) over a long dominant list: reroute.
-        assert!(
-            or_reroute_by_df(LONG, total, 2, 1000),
-            "deep k over a long dominant list should reroute to windowed"
-        );
-        // Shallow k (< rest_df): the rare term still fills the heap, pruning
-        // is alive, stay on MaxScore.
-        assert!(
-            !or_reroute_by_df(LONG, total, 2, 100),
-            "k below the rare term's df keeps pruning alive → MaxScore"
-        );
-        // Exact boundary k == rest_df: the heap needs one doc beyond the
-        // rare term's list, so pruning is already dead → reroute (the test
-        // is `>=`, so the boundary counts).
-        assert!(
-            or_reroute_by_df(LONG, total, 2, RARE as usize),
-            "k exactly at rest_df should reroute"
-        );
-        // One below the boundary (k == rest_df - 1): rare term still fills
-        // the heap, stay on MaxScore.
-        assert!(
-            !or_reroute_by_df(LONG, total, 2, RARE as usize - 1),
-            "k just below rest_df keeps pruning alive → MaxScore"
-        );
-        // Long list but only one term: not an OR.
-        assert!(
-            !or_reroute_by_df(LONG, LONG, 1, 1000),
-            "single term is not a union"
-        );
-        // Small union (dominant list below the floor): too little work to
-        // amortize the window; stay on MaxScore even at deep k. This is the
-        // case that regressed before the floor was added.
-        let small = OR_WINDOWED_MIN_DOMINANT_DF - 1;
-        assert!(
-            !or_reroute_by_df(small, small + RARE, 2, 1_000_000),
-            "a union below the dominant-df floor must not reroute"
-        );
-        // Exactly at the floor with deep k: reroute.
-        assert!(
-            or_reroute_by_df(
-                OR_WINDOWED_MIN_DOMINANT_DF,
-                OR_WINDOWED_MIN_DOMINANT_DF + RARE,
-                2,
-                1000
-            ),
-            "at the dominant-df floor a deep-k union reroutes"
-        );
-    }
-
-    #[tokio::test]
-    async fn windowed_union_negation_agrees_with_bmm() {
-        // The windowed scorer applies the ExcludeFilter (negation) at
-        // drain. Drive a negated query straight through run_windowed_union
-        // and check it matches MaxScore+BMM with the same exclusion — BMM's
-        // negation is the oracle-validated reference, so equality proves
-        // the windowed filter arm. (Calls the scorers directly so the
-        // windowed arm is exercised regardless of the production dispatch.)
-        const N_DOCS: u32 = OR_WINDOW + 1000; // spans more than one window
-        let tok = Arc::new(AsciiLowerTokenizer);
-        let mut b = FtsBuilder::new(tok);
-        b.register_column("body".into(), false).expect("register");
-        for i in 0..N_DOCS {
-            let mut text = String::from("alpha ");
-            if i % 2 == 0 {
-                text.push_str("beta ");
-            }
-            if i % 3 == 0 {
-                text.push_str("gamma ");
-            }
-            if i % 5 == 0 {
-                text.push_str("delta ");
-            }
-            if i % 7 == 0 {
-                text.push_str("epsilon ");
-            }
-            b.add_doc(0, i, text.trim()).expect("add doc");
-        }
-        let blob = Bytes::from(b.finish().expect("finish"));
-        let json = r#"[{"name":"body","tokenizer":"ascii_lower"}]"#;
-        let r = FtsReader::open(blob, json).expect("open");
-        let col = r.resolve_column_id("body").expect("col");
-
-        // (positive terms, negated terms)
-        let cases: &[(&[&str], &[&str])] = &[
-            (&["alpha", "beta", "gamma"], &["delta"]),
-            (&["beta", "gamma", "delta"], &["epsilon"]),
-            (&["alpha", "beta", "gamma", "delta"], &["epsilon", "gamma"]),
-        ];
-        for (pos, neg) in cases {
-            for k in [1usize, 5, 50] {
-                let mut wf = ExcludeFilter::new(
-                    r.build_term_cursors(col, neg, None)
-                        .await
-                        .expect("neg cursors"),
-                );
-                let win = r
-                    .run_windowed_union(
-                        col,
-                        r.build_term_cursors(col, pos, None)
-                            .await
-                            .expect("pos cursors"),
-                        k,
-                        Some(&mut wf),
-                        f32::NEG_INFINITY,
-                        0,
-                        u32::MAX,
-                    )
-                    .expect("windowed");
-                let mut bf = ExcludeFilter::new(
-                    r.build_term_cursors(col, neg, None)
-                        .await
-                        .expect("neg cursors"),
-                );
-                let bmm = r
-                    .run_max_score_bmm(
-                        col,
-                        r.build_term_cursors(col, pos, None)
-                            .await
-                            .expect("pos cursors"),
-                        k,
-                        Some(&mut bf),
-                        f32::NEG_INFINITY,
-                    )
-                    .expect("bmm");
-                assert_eq!(win.len(), bmm.len(), "len {pos:?} -{neg:?} k={k}");
-                for ((dw, sw), (db, sb)) in win.iter().zip(bmm.iter()) {
-                    assert_eq!(
-                        dw, db,
-                        "doc mismatch {pos:?} -{neg:?} k={k}: win={dw} bmm={db}"
-                    );
-                    assert!(
-                        (sw - sb).abs() < 1e-4,
-                        "score mismatch {pos:?} -{neg:?} k={k}: {sw} vs {sb}"
-                    );
-                }
-            }
-        }
-
-        // Sanity: the filter is actually active — at a high k the negated
-        // query must return strictly fewer docs than the positive-only one
-        // (the negated term excludes a non-empty set).
-        let pos: &[&str] = &["alpha", "beta", "gamma"];
-        let neg: &[&str] = &["delta"];
-        let unfiltered = r
-            .run_windowed_union(
-                col,
-                r.build_term_cursors(col, pos, None).await.expect("pos"),
-                N_DOCS as usize,
-                None,
-                f32::NEG_INFINITY,
-                0,
-                u32::MAX,
-            )
-            .expect("unfiltered");
-        let mut f = ExcludeFilter::new(r.build_term_cursors(col, neg, None).await.expect("neg"));
-        let filtered = r
-            .run_windowed_union(
-                col,
-                r.build_term_cursors(col, pos, None).await.expect("pos"),
-                N_DOCS as usize,
-                Some(&mut f),
-                f32::NEG_INFINITY,
-                0,
-                u32::MAX,
-            )
-            .expect("filtered");
-        assert!(
-            filtered.len() < unfiltered.len(),
-            "negation should drop docs: filtered={} unfiltered={}",
-            filtered.len(),
-            unfiltered.len()
-        );
-    }
-
-    #[tokio::test]
-    async fn search_with_algo_empty_and_zero_k_short_circuit() {
-        let (blob, json) = build_blob();
-        let r = FtsReader::open(blob, &json).expect("open");
-        assert!(
-            r.search_with_algo_for_bench("body", &[], 5, OrAlgo::Bmm)
-                .await
-                .expect("empty")
-                .is_empty()
-        );
-        assert!(
-            r.search_with_algo_for_bench("body", &["rust"], 0, OrAlgo::Exhaustive)
-                .await
-                .expect("zero k")
-                .is_empty()
-        );
-    }
-
     #[test]
     fn read_u32_le_and_u64_le_decode_little_endian() {
         let b32 = [0x78, 0x56, 0x34, 0x12];
@@ -3682,16 +2046,6 @@ mod tests {
         scores.insert(9, 5.0);
         let out = top_k(scores, 10);
         assert_eq!(out, vec![(9, 5.0), (5, 2.0)]);
-    }
-
-    #[test]
-    fn drain_top_k_desc_orders_descending_with_tiebreak() {
-        let mut heap: BinaryHeap<TopKEntry> = BinaryHeap::new();
-        heap.push(TopKEntry(1.0, 4));
-        heap.push(TopKEntry(2.0, 1));
-        heap.push(TopKEntry(2.0, 0)); // tie with doc 1
-        let out = drain_top_k_desc(heap);
-        assert_eq!(out, vec![(0, 2.0), (1, 2.0), (4, 1.0)]);
     }
 
     #[tokio::test]
