+++ src/superfile/fts/reader/work.rs
@@ -7,7 +7,7 @@
 //! `fts::reader::*` surface; the summing helpers are `pub(super)`.
 
 use super::cursor::TermCursor;
-use super::phrase::{AnyCursor, PhraseCursor};
+use super::phrase::AnyCursor;
 
 /// Sum of the posting-byte ranges a cursor set indexes into (each cursor's
 /// term metadata + skip table + posting blocks). Feeds the per-query work
